+++ create-manifest.pl
@@ -2,6 +2,13 @@
 use strict;
 use warnings;
 
+# helper function for JSON string escaping
+my %escmap;
+@escmap{split //, "\"\\/\b\f\n\r\t"} = split //, "\"\\/bfnrt";
+sub json_escape ($) {
+	return $_[0] =~ s/([\0-\x1F\\"\x7F-\x{FFFF}])/"\\" . ($escmap{$1} || sprintf "\\u%04X", ord $1)/egr;
+}
+
 # read input from SOUP.user.js if no input files are given on command line
 push @ARGV, grep -f, "SOUP.user.js" unless @ARGV;
 
@@ -18,8 +25,9 @@ while (<>) {
 my %metadata;
 while (<>) {
 	last if m(^// ==/UserScript==\s*$);
-	die "Unrecognized metadata line \"$_\"" unless m(^\s*//\s*\@(\S+)\s+(.*));
-	push @{ $metadata{$1} }, $2;
+	s/\s+$//;  # strip trailing whitespace
+	die "Unrecognized metadata line \"$_\"" unless m(^\s*//\s*\@(\S+)\s*(.*));
+	push @{ $metadata{$1} }, json_escape($2);
 }
 
 # odd-numbered minor versions belong to the devel branch
@@ -29,7 +37,7 @@ $version_name .= " (development)" if (split /\./, $version_name)[1] % 2 == 1;
 # build URL match list for manifest
 my $matches = join ",\n\t\t\t", map qq("$_"), @{ $metadata{match} };
 
-# output manifest (TODO: escape JSON strings properly)
+# output manifest
 print <<"END";
 {
 	"manifest_version": 2,